Add 'level' to 'character_generator'
The 'character_generator' now gives Character instance a level. By 
default, level is equal to 1. If a value other than an integer between 1 
and 20 is passed, the level is validated and either set to the nearest 
integer between 1 and 20 (if the value may be converted to an integer) 
or set to 1 (if the value may not be converted to an integer).